fix(review): separate capability cache integration test
